refactor: post likes and unlikes
