+++ public/javascripts/post-like.js
@@ -2,19 +2,25 @@ const likeSubmitBtn = document.querySelectorAll('.like-submit-btn')
 
 likeSubmitBtn.forEach(item => {
   item.addEventListener('click', async e => {
-    const tweetId = item.dataset.tweetid
-    const likedIcon = item.querySelector('.like')
-    const likedNumbers = item.querySelector('.liked-numbers') || ''
-    if (item.classList.contains('liked')) {
-      likedNumbers.textContent--
-      item.classList.toggle('liked')
-      likedIcon.classList.toggle('user-like')
-      await axios({ method: 'post', url: `${BASE_URL}tweets/${tweetId}/unlike` })
-    } else {
-      likedNumbers.textContent++
-      item.classList.toggle('liked')
-      likedIcon.classList.toggle('user-like')
-      await axios({ method: 'post', url: `${BASE_URL}tweets/${tweetId}/like` })
+    try {
+      const tweetId = item.dataset.tweetid
+      const likedIcon = item.querySelector('.like')
+      const likedNumbers = item.querySelector('.liked-numbers') || ''
+      if (item.classList.contains('liked')) {
+        likedNumbers.textContent--
+        item.classList.toggle('liked')
+        likedIcon.classList.toggle('user-like')
+        await axios({ method: 'post', url: `${BASE_URL}tweets/${tweetId}/unlike` })
+        // await axios.post(`tweets/${tweetId}/unlike`)
+      } else {
+        likedNumbers.textContent++
+        item.classList.toggle('liked')
+        likedIcon.classList.toggle('user-like')
+        await axios({ method: 'post', url: `${BASE_URL}tweets/${tweetId}/like` })
+        // await axios.post(`tweets/${tweetId}/like`)
+      }
+    } catch (err) {
+      console.log(err)
     }
   })
 })
